ripley v0.1.0 Ripley.App

This is the top level app.

This module supervises Ripley.Tally & Ripley.Supervisor, the latter starts one Ripley.Worker for each Ripley.Language, the results from those are sent to Ripley.Tally

Functions

start(type, args)

Used in Ripley.Mixfile.application to start the app, reads the list of languages from config.exs, supervises Ripley.Tally and Ripley.Supervisor

Returns {:ok, pid} or an error, eg {:error, {:already_started, pid}}